gestion de toutes les routes dans un seul fichier
