+++ package.json
@@ -11,6 +11,7 @@
   "license": "ISC",
   "dependencies": {
     "autoprefixer": "^10.4.12",
+    "dotenv": "^16.0.3",
     "ejs": "^3.1.8",
     "express": "^4.18.2",
     "pg": "^8.8.0",
@@ -20,7 +21,6 @@
     "pug": "^3.0.2",
     "tailwindcss": "^3.2.1"
   },
-  "devDependencies": {},
   "repository": {
     "type": "git",
     "url": "git+https://github.com/samuelmbabhazi/AppCRUD.git"
