+++ views/delete.ejs
@@ -4,7 +4,7 @@
 
 <form action="/delete/<%= model.id %>" method="post">
   <div class="form-horizontal">
-    <%- include("_display") -%>
+    <%- include("_editor") -%>
 
     <div class="form-group row">
       <label class="col-form-label col-sm-2"></label>
